refactor: renamed file with functions for working with cached currencies
