+++ src/utils/cachedCurrencies.ts
